Icons

+++ src/components/wallview/Wall/useKeyboardEvents.js
@@ -3,12 +3,12 @@ import { useDispatch } from 'react-redux'
 
 import { deleteArtwork } from '@/lib/features/artistSlice'
 
-export const useKeyboardEvents = (currentArtworkId) => {
+export const useKeyboardEvents = (currentArtworkId, isMouseOver) => {
   const dispatch = useDispatch()
 
   useEffect(() => {
     const handleKeyDown = (e) => {
-      if ((e.key === 'Delete' || e.key === 'Backspace') && currentArtworkId) {
+      if ((e.key === 'Delete' || e.key === 'Backspace') && currentArtworkId && isMouseOver) {
         dispatch(deleteArtwork({ artworkId: currentArtworkId }))
       }
     }
@@ -17,5 +17,5 @@ export const useKeyboardEvents = (currentArtworkId) => {
     return () => {
       window.removeEventListener('keydown', handleKeyDown)
     }
-  }, [currentArtworkId, dispatch])
+  }, [currentArtworkId, isMouseOver, dispatch])
 }
